ai-services/requirements.txt
# Core ML/AI Frameworks
tensorflow>=2.15.0,<2.16.0
torch>=2.1.0,<2.2.0
scikit-learn>=1.3.0,<1.4.0
numpy>=1.24.0,<1.25.0
pandas>=2.1.0,<2.2.0

# Deep Learning and Advanced ML
transformers>=4.35.0,<4.36.0
torch-audio>=2.1.0,<2.2.0
torch-vision>=0.16.0,<0.17.0

# Distributed Computing
ray[default]>=2.8.0,<2.9.0

# API Framework
fastapi>=0.104.0,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0
pydantic>=2.5.0,<2.6.0
PyYAML>=6.0.0,<7.0.0

# Data Processing
scipy>=1.11.0,<1.12.0
statsmodels>=0.14.0,<0.15.0
networkx>=3.2.0,<3.3.0

# NLP and Text Processing
spacy>=3.7.0,<3.8.0
nltk>=3.8.0,<3.9.0

# Experiment Tracking and Model Management
mlflow>=2.8.0,<2.9.0
wandb>=0.16.0,<0.17.0

# Security and Encryption
cryptography>=41.0.0,<42.0.0
pyotp>=2.9.0,<3.0.0

# Database Connectors
psycopg2-binary>=2.9.0,<3.0.0
pymongo>=4.6.0,<4.7.0
redis>=5.0.0,<6.0.0
elasticsearch>=8.10.0,<8.11.0

# Monitoring and Observability
prometheus-client>=0.19.0,<0.20.0
structlog>=23.2.0,<23.3.0
opentelemetry-api>=1.21.0,<1.22.0
opentelemetry-sdk>=1.21.0,<1.22.0

# HTTP and Networking
httpx>=0.25.0,<0.26.0
aiofiles>=23.2.0,<23.3.0

# Configuration and Environment
pydantic-settings>=2.1.0,<2.2.0
python-dotenv>=1.0.0,<1.1.0

# Testing and Development
pytest>=7.4.0,<7.5.0
pytest-asyncio>=0.21.0,<0.22.0
pytest-cov>=4.1.0,<4.2.0
black>=23.11.0,<23.12.0
isort>=5.12.0,<5.13.0
mypy>=1.7.0,<1.8.0

# Utilities
python-multipart>=0.0.6,<0.0.7
click>=8.1.0,<8.2.0
tqdm>=4.66.0,<4.67.0